[bigo] Add extractor
This commit adds a new extractor for https://bigo.tv. Bigo is a live
streaming platform, focused on engagement between broadcasters and
viewers. Viewers can chat & send gifts (using real money) to the
broadcaster. Users primarily use the applications to watch the stream,
but the website is also available.

The extractor uses an internal API which returns JSON. It checks if the
user is currently live and provides an appropriate error message if they
aren't.