Export additional helpful information about livestreams for JSON
It is helpful to know when a live stream or premiere is scheduled, so export "is_upcoming", "live_starttime", and "live_endtime" to the info dict. Also find a non-null value for is_live if possible. is_upcoming is especially useful for downloading youtube live chat, as the chat endpoint may vary.