[afreecatv] Tolerate failure to parse date string
The AfreecaTV extractor currently relies on the fact that video
filenames contain the upload date. Unfortunately, there appear to be
some videos where this is not the case (I would cite examples, but all
that I have encountered have since been deleted or made private). For
these videos, the extractor puts junk data into the upload date field,
which later causes a parse error that fails the download.

This commit adjusts the extractor's logic for parsing the upload date,
so that after it extracts a candidate date value:

1. if the value does not parse as a valid date, the upload date is set
   to None, and
2. if the value *does* parse as a valid date, but the date is outside a
   reasonable range (currently defined as "the 21st century"), the
   upload date is set to None.

I'm open to suggestions on other sanity checks; this commit represents
only the bare minimum necessary to avoid failures on certain videos.

At this time, it seems that a very large majority of video files *do*
have the upload date in their filenames, so the filename heuristic isn't
entirely without merit - i.e. it can probably stay, until and unless
someone implements a more robust way to look up the upload date.