Revert "[downloader/http] Request last data block of exact remaining size"
This reverts commit 6c907eb33ffd79ef9c4761c6460acb31b2eded46.

The use of the Content-Length value here is erroneous and may lead
to truncated downloads if a compression scheme is specified in the
Content-Encoding header, as the Content-Length header refers to the
size of encoded data, not of the raw bytestream. This has been noticed
in the wild with WebVTT subtitle segments.