Fix timestamp overflow adjustment in WebVTT segments
RFC 8216 §3.5 warns us of the possibility of overflow in MPEG
timestamps.  We try to address this warning by checking the MPEG
timestamp value in the X-TIMESTAMP-MAP header against the values found
in earlier segments.  If a segment appears with an MPEG timestamp
lower than those found before, we treat this as overflow condition
and perform an appropriate fixup.  This would have worked well, if it
weren’t for the fact that in some streams empty segments may appear
with a bogus, non-monotone MPEG timestamp.  This has been observed
in the wild.

Perhaps it might be prudent to remove the overflow handling entirely,
given how vanishingly rare this case is expected to be.  It does seem
that the perfect is the enemy of the good.