Modify sanitize_open() to use locked_file(), preventing file corruption when two yt-dlp processes are attempting to write to the same file, where sys.platform != 'win32'.
With this patch, executing a second yt-dlp instance still results in a message: "[download] Resuming download at byte 2159224", however now the the next message is: "ERROR: unable to open for writing: [Errno 11] Resource temporarily unavailable". This fixes the current condition where the two processes would write to the same file, corrupting it. Now the second yt-dlp process moves on to the next required file, so in the case "Downloading 1 format(s): 303+251" if ".f303.webm" is advisory locked by process 1, process 2 will move on to ".f251.webm".

If a .part file is being used (--no-part is not enabled), there is a race condition if the .part file is closed by one process before it is renamed to remove '.part'. In that time window, the other process could modify the .part file before it's renamed. Using --no-part prevents this. It should be possible to lock the .part files.

There is no race condition where process 2 finishes downloading the last needed file (the audio, ".f251.webm" in this example) after skipping the advisory locked video file (".f303.webm") that process 1 is writing to, because process 2 does not attempt to merge the two files together, instead, when process 1 completes it sees the already downloaded audio (*.f251.webm) and merges the two complete files together (usually) into a webm (because it [process 1] encountered no errors).

In the less common 'opposite' case, where process 1 (downloading the video 303) completes while process 2 is still writing the audio 251 file (and therefore has it advisory locked), the 303 process (process 1) throws the message "[download] Resuming download at byte 523264" and then exits with "ERROR: unable to open for writing: [Errno 11] Resource temporarily unavailable" because it is obeying the advisory lock process 2 has on "*.f251.webm". Unfortunately when the 251/audio process (process 2) completes, it does not attempt to merge the now complete video and audio files, since it has no way to know if the 303 file is complete (and it [process 2] already encountered an error trying to write to the 303 video file). To correct this, process 2 could check if another process still holds an advisory lock on the 303 file, and if not, obtain a lock on it and then compare the 303 file stat() with the expected metadata. This corner case is still left to fix. Re-executing a third yt-dlp process to the same target with the completed audio and video files (still) results in the expected merge of the two.

Untested on Windows, locking is not implemented, should be a unchanged if the included function signature modification is correct.

On other platforms there should be no case that results in a corrupt output file.