[Coub] Fix media format identification
The existing code incorrectly labelled some formats as containing both
audio and video. While it's true that these formats may include *an*
audio stream in them, this is not the same thing as the "main" audio
soundtrack for the Coub itself; this main audio is only delivered via
a separate, audio-only format.

Also, add "higher" as a recognized quality string, considered better
than "high". Without this, the extractor correctly lists the "higher"
quality video, but it is considered the *worst* format rather than the
best.