[ceskatelevize] Update to 2021-12-06 changes (#1899)
After a site redesign on 2021-12-06, player iframe is no longer present in
the page from the start and is only inserted after pressing the "play"
button. We need to query /v-api/iframe-hash/ to get hash to pass to
iFramePlayer.php and extract "IDEC" id from a json in the original page.