Fix parsing YYYYMMDD dates in Nov/Dec
Surprisingly, the date format %Y%m%d%H%M will successfully match against
one-digit month, day, hour, and minute strings, even though %m et al.
are documented as being zero-padded. For most of the year, one of the
digits in a two-digit month is zero, which prevents the month from being
misinterpreted as month and date. However, for dates in November, 11 can
parse as January 1st (similarly, 12 as January 2nd). In this case, the
actual day will be interpreted as an hour and a minute.

This commit adds a format string of %Y%m%d to our supported date format
strings. Because unified_strdate doesn't exit its parsing loop on the
first successful match, the *bottom* of the list represents formats with
higher priority, so we add this string directly below its problematic
relatives.