Update mildom.py
Added an extra test case for MildomUserVod, and a few for MildomVod. May have gone a bit overboard, but the original test for the user playlist alone did not show any errors for the original bug (apparently it just counts the playlist entries, but does not validate that the videos actually exist).