[viki] Cache session token
If you log in too many times in a short time period, Viki will start
throwing a 429 Too Many Requests error and prevent logins to that
account from any IP for a while (may be even around 30 minutes).

This will prevent that from happening by caching the session token.
Viki doesn't seem to provide any info on when it expires, so login
is retried if the API throws an invalid token error.