Rename `_login` to `_perform_login`
`_login` is a commonly used name. If an extractor with a `_login` method
defined with wrong signature is accidentally merged (from upstream),
the extractor will break. Even worse, if it happens to have the right
signature, the issue may go unnoticed.

Any merge conflict that may arise due to this renaming is worth it to avoid such mistakes