[general options] add --force-extractor
This commit adds support for a new command-line flag, --force-extractor,
that takes as an argument the name of an extractor. This extractor is
pased to extract_info as the ie_key, which causes the given extractor to
be used directly without having to run __init__ on all supported
extractors.

This is important because each extractor typically needs to compile a
regex, and all of that regex compilation adds up. In a profiler, this
extractor initialization occupies 80% of startup time on my test
machine (the rest is module import and checking ffmpeg versions).

Passing --force-extractor therefore gives the user a faster startup when
a given invocation is known to always need a particular, named
extractor.