[nebula] Add support for subscriptions
Nebula has a feed of all episodes from channels/collections the user is
following, called `myshows` or `My Videos` in URLs and UI respectively.
This is the equivalent to YouTube's `Subscriptions` page.

This change adds support for downloading this feed via the
`/myshows` URL.

It also fixes extractor tests and replaces the call to
`_get_login_info()` with the new `_perform_login()` system.