motherless: update group and gallery url pattern
The abstraction does not yet handle these variations.

For galleries we need to paginate on /GV<ID>?page=n, but ideally we
could also just accept /G<ID> in addition to /GV<ID>.

Same goes for groups, where /g/<NAME> refers to the group and /gv/<NAME>
for the groups videos, where the pagination happens.