turn cookie store detection on it's feet
All the Unix-like platforms (Linux, *BSD) use the same location for
the browser profile directories, so it makes some sense to make
that the default instead of tracking all the platform names; and
treat Windows and OSX as the exceptions they are (based on the number
of platform names, not installs). This is especially useful as the
"platform" name on at least FreeBSD changes with each major release
(e.g. freebsd12, freebsd13, ...) and would have to be updated for
every release without this change here.
At the same time, fix the error condition: the question is not whether
we are on a "supported" platform but if the browser profile directory
exists in the place where we expect it. That handles all the cases like
"browser not available", "browser not installed", "profile not
initialized (yet)" and "non-standard profile dir location". Change
error message to match that behaviour.