Make the build reproducible
Whilst working on the Reproducible Builds effort [0] I noticed that
yt-dlp could not be built reproducibly.

This is due to `sort_ies` iterating over a Python set() data structure
in a nondeterminstic order. This patch sorts the type objects according
to their name.

I originally filed this in Debian as bug #1014041 [1].

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/1014041