[aes] Add key expansion test
I had a hard time using the 'key_expansion' function, because although
it documents its inputs and outputs; it's not really clear on which kind
of input it really takes.

For example, int[] could mean [91324035, 135346496] (array of 32-bit
integers) or [12, 75, 255] (array of 8-bit integers), just after you
have spent some time on this, you understand that it means an array of
8-bit integers.

Having this test will give us at least a clue.