Update YoutubeDL.py
If the info_dict contains an extension, save it as the preferred format for merge (if not specified otherwise). In case of a merge, the extension will be overridden by the chosen (best) format without notice (for backwards compatibility, as the comment says).

This looks very ugly but still safer than overriding extensions in all available formats. It would probably be much better to leave the info_dict alone.