Add devscript to generate regexes for known Invidious and Piped sites
Both Invidious and Piped have an index of known-good instances. Parse
those lists, match them against the list of instances in the
YoutubeBaseInfoExtractor code, and generate regexes for all instances
that don't match existing patterns.

The same script can also be used to add single instances via the command
line which are not in the official indexes.