[utils] Don't trust Content-length when Content-encoding is present
Our YoutubeDLHandler is doing auto-decompression, so that we cannot
compare Content-length to the size of extracted data. Although HttpFD has
specified Youtubedl-no-compression, the web server may still ignore this
and return deflated data (as the situation in #3772).

This change attempts to retain the Content-encoding header for later
comparison in HttpFD so that we could discard data_len when the web
server returns compressed data. I didn't find the reason why this header
was previously removed but didn't yet find a counter example either.

Fixes #3772 in my experiments.