[extractor/twitter] fix extraction through rest api on links to retweet id
a retweet has its own id. with recent lack of quality assurance,
the "[someone] liked your retweet" notifications are now pointing
to the id of the retweet, instead of id of the post.

with this change, extractor looks for the retweeted tweet, which
is contained inside the response under "retweeted_tweet" key.