[downloader/fragment] Partially revert 4823ec9f461512daa1b8ab362893bb86a6320b26
Currently, when `--no-continue` is used, yt-dlp will attempt to download
from the first fragment and **append** decrypted fragments to the temp
file (or the actual video file), causing data duplication when resuming
a partially downloaded video.

This fixes the issue by partially reverting the behavior of the fragment
downloader to match its description in the README file.