__init__.py: remove unnecessary f-string
When the python runtime version is < 3.7, an Exception is raised. The message
contained in that version contained an f-string with no placeholders.

This triggered a warning in pyflakes, that was subsequently silenced with a
F541 directive.

We can simply use a plain string and get rid of the "noqa" directive.