Update yt_dlp/extractor/err.py - let Python throw error on unexpected input
I said it should be fatal. As in it should not be caught.
There is nothing logical we can do without the data, we don't need to prevent the error to then immediately throw a new error due no data.
We also don't expect the API to ever not return this data and if it does the extractor will most likely need to be updated.

Co-authored-by: sepro <4618135+seproDev@users.noreply.github.com>