[ie/NhkRadiru] Extract extended description
Pieces that make up an extended description are available through a
programme info API. They are, in order:
- short summary
- longer description
- the host(s)
- list of songs played

The host(s) field is free text, so it wouldn't be practical to put it in
`cast`. the order of items is copied from the mobile app - other orders
are used on the website's timetable, but I couldn't find the API etc
where they actually come from.

API info is gotten from https://www.nhk.or.jp/radio/config/config_web.xml,
which is the same endpoint used in NhkRadiruLiveIE. I don't think a
BaseIE etc is worth it at this time.

It would probably be good to cache that info, or at least only download 
it once when downloading a playlist, but I'm not sure how to go about it.