[AltCensoredChannelIE] Fix page_count regular expression
The current regexp does not take into account that channel_id may contain
one or more dash "-" character(s). If such a channel is downloaded, the
detection of the number of pages fails and the default value of "1" is used.
This results in only the first page being downloaded.

A ChannelIE download test was added to cover the scenario above. Also fixed
a couple of failing existing tests in this extractor.

Signed-off-by: Marc dumais <marc.dumais@protonmail.com>