[ie/youtube:tab] Fix the first 2 test cases
The first 2 test cases (`test_YoutubeTab` and
`test_YoutubeTab_1`) also demonstrate the issue
with incorrectly split tag names
(`["\"foo", "bar\""]` instead of `["foo bar"]`).
I fixed them for the new behaviour. To make them
pass, I also had to update other fields in the
test data (it appears since this test case was
written, the channel has changed names, and the
description has changed).

I did not go any further fixing test cases in this
commit, because many of them seemed to be
out-of-date and not passing even before this fix.