CNNArticleIE Extractor: Update to recognize modern video links
CNNArticleIE currently matches with URLs but fails to extract video.
Update CNNArticleIE to work with the way most CNN video
links are embedded, and update the tests to include some of these
links. The type of URL for this can just use the default extractor.
Also updating the regex to capture a date or text before
the /video/ subcategory, as many video links are structured
this way. Removing old test that has an old article without
available media anymore.

Example URLs:

Landing pages:
https://www.cnn.com/videos
https://www.cnn.com/videos/
https://edition.cnn.com/videos

Specific Videos:
https://www.cnn.com/2024/05/31/sport/video/jadon-sancho-borussia-dortmund-champions-league-exclusive-spt-intl
https://edition.cnn.com/2024/06/11/politics/video/inmates-vote-jail-nevada-murray-dnt-ac360-digvid
https://www.cnn.com/2024/06/11/style/video/king-charles-portrait-vandalized-activists-foster-intl-digvid