[globo] Use new server in new domain
Domain has changed from `globo-ab.globo.com` to `ab.g.globo`.
Server in the old domain is no longer responding to HTTPS requests.

As of 2024-12 the extractor is still broken as it will fail down the
line due to API changes in Globo side, but this commit is one step
towards fixing it.