[downloader/hlsnative] Fix byterange logic
In cases where the EXT-X-MAP tag had the BYTERANGE attribute set it would
set the byte_range variable in a too-broad scope.

If subsequent fragments didn't have their own byterange specified
(for example because they had unique URIs or the byterange was in the URI)
the fragment would erroneously limit its byterange to that of the MAP.

This commit decouples the MAP byterange from Media Segment byteranges
and makes Media Segment byteranges picked up from a previous EXT-X-BYTERANGE
tag follow the RFC8216 spec by only using it on the next Media Segment.