Update msn.py
Testing to see if my msn_video_downloader.py integration works comfortably within the yt-dlp environment without causing error or conflict. This is a working progress and not yet finished until it can be tested. I really hope this is not attempting to replace the actual "master" branch and that this is just my own fork being edited as I have never done this before.