Update _utils.py
Remove CLOCK$ as it is a valid filename in Win10. It was reserved in NT or older systems as referenced here: https://answers.microsoft.com/en-us/windows/forum/all/folder-names/2f9bb53a-da99-45e6-90f7-50d1399842aa. Testing on Win10 VM also shows that it's possible to create files/directories with it as a name.