[fd/hls] Fix `format_index=0` parsing with `--hls-split-discontinuity`
When treating each discontinuity as a separate format, this code should
only consider items within the requested discontinuity. But when the
requested discontinuity has index 0, it currently considers everything
because 0 is not truthy.

Fix the issue by explicitly checking for non-`None` `format_index`,
which indicates that `--hls-split-discontinuity` was passed.

Fixes: 310c2ed2c648 ("Better support HLS media discontinuity and fully support media initialization (#105)")