Remove dead-code assertion for `playlist_duration_sum`
There are no current usages, which can be verified with
    git grep playlist_duration_sum

The last usage was removed in 92dcba1e1cc411eecb64792e78ad8f3b125691c8,
as per
    git log -S playlist_duration_sum -p

That commit was made over 10 years ago and it seems unlikely than a test
inside a plugin would make use of this special-case assertion. So I
think it is worth removing.