fix(parser): Adapt to html.parser changes in Python 3.14
Python 3.14's html.parser is now more compliant with the HTML5
standard and no longer ignores whitespace around the equals sign (=) in
attributes. This change caused `extract_attributes` to fail when parsing
tags with attributes like `<e x= y>`.

This commit restores the previous behavior by pre-processing the HTML
string with a regex to remove this whitespace before parsing. The fix is
conditional to only apply to Python 3.14 and newer.

Ref: https://github.com/python/cpython/issues/135661

Signed-off-by: Mikel Olasagasti Uranga <mikel@olasagasti.info>