[mgtv] Use a different URL when cookies are present.
Previously, the mgtv extractor used the same URL when cookies are
present vs active. Unfortunately, that URL always downloads only the
first five minutes of a video, even when cookies would allow viewing the
entire video.
This commit updates the extractor to use a different URL when cookies
are offered, to allow downloading full videos when the proper
credentials are given in cookies.