fix(arte): add required property url
The info dict documentation lists url as a required property:
```
url:            Final video URL.
```
https://github.com/yt-dlp/yt-dlp/blob/master/yt_dlp/extractor/common.py#L281

The extractor should also return this property.

Initially, I thought `webpage_url` would not be in the documentation,
so I renamed `webpage_url` to `url`.
Later, I learnt that there is also the optional property `webpage_url`.
This is why I'm adding the required url and keeping `webpage_url` as is.